refactor: reworked variables
Before variables declared via the ':'-operator were all arrays, which is
bad design because i shot myself in the foot while trying to implement
and, or and arithmetic operations. Now variables defined with one child,
such as the following are no longer an arrays, which was the default
behaviour before.

    (: a 1)     ;; <- a just contains 1
    (: b 1 2 3) ;; <- b contains an array with 1,2 and 3

Due to this change i could slim down the following expressions
significantly:

- add           (+)
- sub           (-)
- div           (/)
- mul           (*)
- mod           (%)
- and           (&)
- or            (|)
- equals        (=)

Which does however mean the above do no longer accept variables that
are arrays. In the future a function to accumulate values of an array
and one to iterate over an array are planned.

The change did also allow me to inline 'expr.isType' into its now only
usage in 'expr.castPanicIfNotType'.

Changelist:
- removed arrays support for most operators
- updated tests to match the above changes
- moved from all variables are arrays to declare variable with just one
  argument -> normal variable, declare variable with multiple arguments
  -> array