feat: for loop
This commits adds support for iterating over containers with the
following syntax:

    (let sum 0)
    (let arr 1 2 3 4 5 6 7 8 9)
    (for
        (_ e)
        arr
        (let sum
            (+ e sum)))
    (put sum)

The above calculates the sum of the elements in the 'arr' container. The
first argument of the loop defines the parameter which can be used in
the scope of the loop. The second argument defines the container to
iterate over. Any following statement will be executed for every item in
the container 'arr'.

Change list:
- added for loop support
- added for loop test
- added expr.For
- added token.FOR
- renamed repl commands:
    symbols -> syms
    funcs -> funs