feat: add support for array creating via spread syntax (x..y)
Creating an array was a cumbersome endeavour due to the need to
explicitly specify all elements of said array. This commit aims to
circumvent this by introducing a new syntax for creating large number
arrays, which is particularly useful for iterating over large indexes.

The new syntax can be used as shown below:

    (let a 1..10)

This creates an array containing all integers from 1 up until and
including 10. Inspecting the symbol table via '~syms' in the repl will yield:

    ß :: ~syms
    map[string]interface {}{"a":[]interface {}{1, 2, 3, 4, 5, 6, 7, 8, 9, 10}}

Semantics:

    1..10 <- upper bound
    ^
    |
    lower bound

The upper bound is always required, trying to use '1..' or '..' will
result in a runtime error. The lower bound can be omitted, if so the
interpreter assumes the lower bound to be 0.

    (let a ..10)

This includes all numbers from 0 up to and including 10:

    ß :: ~syms
    map[string]interface {}{"a":[]interface {}{0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10}}

Changes:
- add array spread syntax
- fixed runtime panic on variable without arguments: (let a) failed
  before
- add array spread syntax to examples
- add planned match operator feature