core/lexer: remove '.' as valid float postfix
The following is from now on no longer lexed as TOKEN_FLOAT, but instead as TOKEN_FLOAT and TOKEN_DOT: '0.'. This change was made to support indexing into arrays and objects such as '[arr.0.name]' in which the lexer would result in TOKEN_IDENT, TOKEN_DOT, TOKEN_FLOAT and TOKEN_IDENT, thus swallowing the dot behind the index of the array. '.' between two integers are still perfectly valid: '0.0', '1.2'