core: support for arrays as primitives
This commit introduces a new way of creating arrays inside of objects
and as return values for functions. Previously arrays could only be
created as parts of 'let' expressions. Due to the previous changes to
index notation '[]' are now free to be reused by the syntax for array
creation:

    (let a 1 2 3 4 5)
    ;; equal to
    (let a [1 2 3 4 5])

    ;; in objects:
    (let person {
        name: "anon"
        aliases: ["anon1" "unon2" { f: "aa" s: "non" }]
    })

    (put person.aliases.2.f person.aliases.2.s)

Arrays can contain any other type, such as: float, string, bool,
objects and other arrays.

Changes:
- core/expr: Array expression
- core/parser: Parser.parseArray method
- core/run: log all non sophia specific errors
- core/serror: Panic prefixed with 'sophia: '