core/parser: new array declaration
The new array declaration notation is using '#' as the operator and can
include strings, booleans, floats, objects, the second notation is and
will be further available:

    (let array (# 1 2 3 4 5))
    ;; equal
    (let array 1 2 3 4 5)

The previous syntax for declaring arrays, see below, used the
[<element...>] notation.

    (let array [1 2 3 4])

However i want to move from target.index/field to target[index/field]
for object and array indexing. The issue with the aforementioned being
the lexer can not distinguish between 'array.2.0' and 'array. 2.0' and
therefore treats the 2.0 as one, when he should treat it as 'array . 2 .
0', this warrants the change to array index notation from 'array.index' to
'array[index]'. However there would be an inconsistency between indexing
array elements and object fields: 'array[index]' and 'object.field'. To
support dynamic object field index and keep the index notation
consistent i have decided on removing the usage of [] for array
declaration and reserving it for the future changes i previously
explained.